+++ reference/cs/ec-cines/EcCines/Scripts/Index.js
@@ -1,6 +1,10 @@
 ﻿windowOpen = function (movieId) {
-    var url = window.location.origin + "/moviedetailiframe.aspx?m=" + movieId;
-    window.open(url, "VistaPreviaPelicula");    
+    var url = window.location.origin;
+    if (url == "undefined") {
+        url = window.location.protocol + "//" + window.location.host;
+    }
+    url = url + "/admin/moviedetailiframe.aspx?m=" + movieId;
+    window.open(url, "VistaPreviaPelicula");
     return false;
 }
 $(document).ready(function () {